requirements.txt
deepgram-sdk==4.7.0
pydub
google-generativeai
langchain-google-genai
python-dotenv
pandas
sendgrid
openpyxl
streamlit
fastapi
uvicorn[standard]
python-multipart
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
PyJWT==2.8.0